refacto: reorganize custom types
