+++ nlptest/utils/custom_types/helpers.py
@@ -1,0 +1,47 @@
+from pydantic import BaseModel
+
+
+class Span(BaseModel):
+    """Representation of a text's slice"""
+    start: int
+    end: int
+    word: str
+
+    def shift_start(self, offset: int) -> None:
+        """"""
+        self.start -= offset
+
+    def shift_end(self, offset: int) -> None:
+        """"""
+        self.end -= offset
+
+    def shift(self, offset: int) -> None:
+        """"""
+        self.start -= offset
+        self.end -= offset
+
+    def __hash__(self):
+        """"""
+        return hash(self.__repr__())
+
+    def __eq__(self, other):
+        """"""
+        return self.start == other.start and self.end == other.end
+
+    def __str__(self):
+        """"""
+        return f"<Span(start={self.start}, end={self.end}, word='{self.word}')>"
+
+    def __repr__(self):
+        """"""
+        return f"<Span(start={self.start}, end={self.end}, word='{self.word}')>"
+
+
+class Transformation(BaseModel):
+    """
+    Helper object keeping track of an alteration performed on a piece of text.
+    It holds information about how a given span was transformed into another one
+    """
+    original_span: Span
+    new_span: Span
+    ignore: bool = False
